certquest-api-main/pyproject.toml
[project]
name = "certquest-api-main"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "azure-cosmos>=4.9.0",
    "azure-identity>=1.23.0",
    "fastapi[standard]>=0.115.12",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
]
